Implement simple calculator
Simple techniques were used to build a calculator which uses only
html and css. The intention is to have to calculator work without
javascript and then progressively enhance it. A simple css
styling was applied to the html to make it look okay.
The formula used on the server is just an example formula for
demonstration purpose. It is not meant to model an annuity in any
way.