Add client side dynamic calculation
Implemented client side javascript to perform the same calculation
as is done on the server side. The implementation of the calculation
is *not* shared between the client code and the server code. This
should be the next step. Some of the styling could be improved.
Perhaps a grid layout would be better than a flex layout for the
form. An empty canvas element is present at the end of the document.
This is where a chart.js chart will be rendered.