Share logic between client and server
There is some logic which is used by both the client and the server
which is identical. It is best to only maintain one implementation
of the logic. To do this webpack was used. The current directory
structure feels like it could be improved to make things clearer,
but it works for now.

Webpack has been set up to transcompile es2015 to es5 (or earlier?)
so that we don't have to worry about browser compatibility when
writing the shared logic code.