Setup simple build system with sass
I want to be able to use sass to produce the style sheets.
Express is able to process sass on the fly using the sass
middleware. However this adds an un-nessesary to the express
app. All I need is for it to be processed at build time.
To get this done, the node-sass package was used. The
Libsass seems to come precompiled so it shouldn't be a hassle.
Simple npm scripts were set up so that when npm run build is
run it will run webpack and node-sass. This build script solution
seems like it will be tiresome to manage. I need to look into
something like gulp.

Note: This change did not port the css to scss. Since scss is a
superset of css, it should run just fine for now.