Set up gulp build script
For some things the npm scripts were cumbersome and slow.
The task of building the css and client side javascript has been
implemented in a gulp script.

The gulp build script does two tasks. It processes the styles and
the client side javascript. To process the styles, a libsass gulp
wrapper is used. The css is then further run through
autoprefixer for automatic compatibility with old browsers and
then through cssnano to compress it. The client side javascript
is processed using webpack. The client side javascript is then
compressed.

Source mappings are produced for both the css and client side
javascript. The seem to work well in the Chrome developer tools.

The npm scripts is still used to call gulp build and then to
start the express server because it is simpler that way.

Possible improvements to the tasks is to implement a more convinient
development watcher. This would have to a few things:
* watch scss files and rebuild them when changed.
* watch client side js (and shared js) and rebuilt when changed.
* Restart the express server when server side and shared js is changed.

Another thought is that it would be nice to have a server log
printed to the terminal when running the server. Much like the
default behaviour provided by the express-generator tool.