Move webpack config to external script
Moved the webpack config to its own config file just as if it was
run from the webpack cli. Some parts of the config are only set
up in the gulp script. The intention of this is to have the
nessesary webpack config in the webpack.config.js and then have the
build type specific config set up the in appropriate build script.
For example if I may decide that I want a build task which doesn't
specify the debug flag or output a sourcemap. However, not running
Babel seems unlikely.

In the spirit of this change, the .min suffixes were dropped from
the processed style sheet and script bundle. There is no benefit
from using this naming convention. Even if UglifyJs and cssnano
were not used, they are still not the original source code.