Refine task scripts
The gulp task scripts were refined. All tasks seem to work
well. Gulp injection doesn't seem to work, but that is not
a problem.

A serve script was set up in bin/serve.js. This script launches
a server which serves the server.js app, serves the static files
in dist and does logging using morgan. For production, this
script may have to be changed a bit. E.g. we would probably have
a nginx proxy or something similar which provided all the
functionality of the serve.js script. All serve.js would have to
do is launch the app provided by server.js.

In the gulpfile.js, the port number used by serve.js is assumed.
I wonder how the port number might be evaluated?