Update npm scripts
Updated existing npm scripts to run the new gulp tasks.
Added new scripts to run the linting commands.
Unfortunately, npm run outputs a distracting error message when
one of the linters exits with an error code. The linters exit with
an error code when ever they come across something that doesn't
pass the test so it is not unexpected.